refactor: 앨범 상세 로그 피드 컴포넌트 분리 #149
