+++ VibeTrip/Features/Album/Views/AlbumDetailView.swift
@@ -1042,29 +1042,6 @@ private struct AlbumMenuItemButtonStyle: ButtonStyle {
     }
 }
 
-// MARK: - LogMenuItemButtonStyle
-
-private struct LogMenuItemButtonStyle: ButtonStyle {
-    
-    private enum Constants {
-        static let horizontalPadding: CGFloat = 16
-        static let verticalPadding: CGFloat = 8
-        static let cornerRadius: CGFloat = 8
-        static let highlightBackground = Color(red: 0.92, green: 0.92, blue: 0.98)
-        static let animationDuration: Double = 0.1
-    }
-    
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .padding(.horizontal, Constants.horizontalPadding)
-            .padding(.vertical, Constants.verticalPadding)
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(configuration.isPressed ? Constants.highlightBackground : Color.clear)
-            .cornerRadius(Constants.cornerRadius)
-            .animation(.easeInOut(duration: Constants.animationDuration), value: configuration.isPressed)
-    }
-}
-
 // MARK: - AlbumDetailAlbumMenuPopup
 
 private struct AlbumDetailAlbumMenuPopup: View {
@@ -1130,486 +1107,6 @@ private struct AlbumDetailAlbumMenuPopup: View {
         }
         .disabled(isDisabled)
         .buttonStyle(AlbumMenuItemButtonStyle())
-    }
-}
-
-// MARK: - AlbumDetailLogMenuPopup
-
-private struct AlbumDetailLogMenuPopup: View {
-    
-    private enum Constants {
-        static let popupWidth: CGFloat = 140
-        static let padding: CGFloat = 8
-        static let itemSpacing: CGFloat = 8
-        static let cornerRadius: CGFloat = 12
-        static let itemFontSize: CGFloat = 14
-    }
-    
-    let onEditLog: () -> Void
-    let onDeleteLog: () -> Void
-    
-    var body: some View {
-        VStack(alignment: .center, spacing: Constants.itemSpacing) {
-            Button(action: onEditLog) {
-                Text("로그 수정")
-                    .font(.setPretendard(weight: .medium, size: Constants.itemFontSize))
-                    .foregroundStyle(Color.text)
-            }
-            .buttonStyle(LogMenuItemButtonStyle())
-            
-            Button(action: onDeleteLog) {
-                Text("로그 삭제")
-                    .font(.setPretendard(weight: .medium, size: Constants.itemFontSize))
-                    .foregroundStyle(Color.text)
-            }
-            .buttonStyle(LogMenuItemButtonStyle())
-        }
-        .padding(Constants.padding)
-        .frame(width: Constants.popupWidth, alignment: .center)
-        .background(.white)
-        .cornerRadius(Constants.cornerRadius)
-        // 로그메뉴팝업 -> detailMenu shadow
-        .appShadow(.detailMenu)
-    }
-}
-
-// 가림 판단용: 콘텐츠 좌표계 이름 + 카드 하단 위치 보관
-private let albumDetailScrollSpace = "albumDetailScroll"
-
-private final class CardBottomTracker {
-    var contentMaxY: CGFloat = 0   // 카드 하단의 콘텐츠 좌표계 기준 y
-}
-
-// MARK: - AlbumDetailLogFeedSection
-
-private struct AlbumDetailLogFeedSection: View {
-
-    @ObservedObject var viewModel: AlbumDetailViewModel
-    let onEdit: (AlbumLogEntry) -> Void
-    let onRevealCard: (Int, CGFloat) -> Void // (카드 id, 카드 하단 콘텐츠 y) — 가릴 때만 스크롤
-    
-    private enum Constants {
-        static let horizontalPadding: CGFloat = 20
-        static let topPadding: CGFloat = 4
-    }
-    
-    var body: some View {
-        let groups = viewModel.groupedLogs
-        let lastEntryId = viewModel.logs.last?.id
-        
-        VStack(alignment: .leading, spacing: 20) {
-            ForEach(groups.indices, id: \.self) { groupIndex in
-                AlbumDetailLogDateGroup(
-                    label: groups[groupIndex].dateLabel,
-                    entries: groups[groupIndex].logs,
-                    lastEntryId: lastEntryId,
-                    onLastAppear: {
-                        guard let id = lastEntryId else { return }
-                        await viewModel.loadMoreIfNeeded(lastId: id)
-                    },
-                    onDeleteLog: { logId in
-                        viewModel.requestDeleteLog(id: logId)
-                    },
-                    onEdit: onEdit,
-                    awaitingImageLogIds: viewModel.awaitingImageLogIds,
-                    onRevealCard: onRevealCard
-                )
-            }
-        }
-        .padding(.horizontal, Constants.horizontalPadding)
-        .padding(.top, Constants.topPadding)
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-}
-
-// MARK: - AlbumDetailLogDateGroup
-// 날짜별 로그 그룹 (헤더 + 카드 목록)
-
-private struct AlbumDetailLogDateGroup: View {
-    let label: String
-    let entries: [AlbumLogEntry]
-    let lastEntryId: Int?
-    let onLastAppear: (() async -> Void)?
-    let onDeleteLog: (Int) -> Void
-    let onEdit: (AlbumLogEntry) -> Void
-    let awaitingImageLogIds: Set<Int>
-    let onRevealCard: (Int, CGFloat) -> Void
-
-    private enum Constants {
-        /// 로그 카드 간 간격
-        static let itemSpacing: CGFloat = 20
-    }
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: Constants.itemSpacing) {
-            ForEach(entries) { entry in
-                AlbumDetailLogItemCard(
-                    entry: entry,
-                    isLast: entry.id == lastEntryId,
-                    onLastAppear: onLastAppear,
-                    onDeleteLog: onDeleteLog,
-                    onEdit: onEdit,
-                    showSkeletonIfNoImages: awaitingImageLogIds.contains(entry.id),
-                    onRevealCard: onRevealCard
-                )
-                .id(entry.id)   // 펼침 시 이 카드로 scrollTo 하기 위한 앵커
-            }
-        }
-    }
-}
-
-// MARK: - AlbumDetailLogItemCard
-// 개별 로그 아이템 카드
-
-private struct AlbumDetailLogItemCard: View {
-    let entry: AlbumLogEntry
-    let isLast: Bool
-    let onLastAppear: (() async -> Void)?
-    let onDeleteLog: (Int) -> Void
-    let onEdit: (AlbumLogEntry) -> Void
-    // 방금 저장한 로그처럼 이미지 응답이 비어 있을 때 슬라이더 자리에 스켈레톤을 그릴지 여부
-    let showSkeletonIfNoImages: Bool
-    let onRevealCard: (Int, CGFloat) -> Void
-
-    // 로그 옵션 팝업 표시 여부
-    @State private var isMenuVisible: Bool = false
-    // 카드 하단의 콘텐츠 좌표계 기준 y (가림 판단용)
-    @State private var bottomTracker = CardBottomTracker()
-
-    private enum Constants {
-        static let dateFontSize: CGFloat = 14
-        static let menuIconSize: CGFloat = 16
-        static let menuTouchTarget: CGFloat = 44
-        static let dateToImageSpacing: CGFloat = 8
-        static let contentSpacing: CGFloat = 12
-        static let menuAnimationDuration: Double = 0.15
-        static let menuTopOffset: CGFloat = 26
-        static let menuTrailingPadding: CGFloat = 17
-        static let labelColor = Color(red: 0.74, green: 0.75, blue: 0.76)
-        // 펼침 애니메이션이 끝나 카드 높이가 확정된 뒤 스크롤하도록 약간 지연
-        static let expandScrollDelay: Double = 0.3
-    }
-    
-    /// postedAt ISO8601 → "yyyy년 M월 d일" 포맷
-    private var dateLabel: String {
-        let displayFormatter = DateFormatter()
-        displayFormatter.locale = Locale(identifier: "ko_KR")
-        displayFormatter.dateFormat = "yyyy년 M월 d일"
-        guard let date = AlbumDetailViewModel.parseISO8601Date(entry.postedAt) else {
-            return entry.postedAt
-        }
-        return displayFormatter.string(from: date)
-    }
-    
-    var body: some View {
-        ZStack(alignment: .topTrailing) {
-            VStack(alignment: .leading, spacing: 0) {
-                // 날짜 + 로그 옵션 버튼
-                HStack {
-                    Text(dateLabel)
-                        .font(.setPretendard(weight: .medium, size: Constants.dateFontSize))
-                        .foregroundStyle(Constants.labelColor)
-                    Spacer()
-                    Button {
-                        withAnimation(.easeInOut(duration: Constants.menuAnimationDuration)) {
-                            isMenuVisible.toggle()
-                        }
-                    } label: {
-                        Image(systemName: "ellipsis")
-                            .font(.system(size: Constants.menuIconSize))
-                            .foregroundStyle(Constants.labelColor)
-                            .frame(
-                                width: Constants.menuTouchTarget,
-                                height: Constants.menuTouchTarget
-                            )
-                    }
-                    .buttonStyle(.plain)
-                    .padding(.trailing, -4)
-                }
-                
-                // 이미지 슬라이더: 이미지 있을 때 본 슬라이더, 응답 대기 중인 로그는 스켈레톤
-                if !entry.images.isEmpty {
-                    AlbumDetailLogImageSlider(images: entry.images)
-                        .padding(.top, Constants.dateToImageSpacing)
-                } else if showSkeletonIfNoImages {
-                    AlbumDetailLogImageSkeleton()
-                        .padding(.top, Constants.dateToImageSpacing)
-                }
-                
-                // 텍스트 + 더보기/접기
-                AlbumDetailLogTextSection(text: entry.description, onExpand: {
-                    // 펼침 레이아웃이 잡힌 뒤, 카드 하단 위치와 함께 자동 스크롤 요청
-                    DispatchQueue.main.asyncAfter(deadline: .now() + Constants.expandScrollDelay) {
-                        onRevealCard(entry.id, bottomTracker.contentMaxY)
-                    }
-                })
-                .padding(.top, Constants.contentSpacing)
-            }
-            
-            // 팝업 표시 시: 외부 탭 dismiss 영역 + 팝업
-            if isMenuVisible {
-                // 카드 범위 내 외부 탭 감지
-                Color.clear
-                    .contentShape(Rectangle())
-                    .onTapGesture {
-                        withAnimation(.easeInOut(duration: Constants.menuAnimationDuration)) {
-                            isMenuVisible = false
-                        }
-                    }
-                
-                // 수정 및 삭제 팝업
-                AlbumDetailLogMenuPopup(
-                    onEditLog: {
-                        isMenuVisible = false
-                        onEdit(entry)
-                    },
-                    onDeleteLog: {
-                        withAnimation(.easeInOut(duration: Constants.menuAnimationDuration)) {
-                            isMenuVisible = false
-                        }
-                        onDeleteLog(entry.id)
-                    }
-                )
-                .padding(.top, Constants.menuTopOffset)
-                .padding(.trailing, Constants.menuTrailingPadding)
-            }
-        }
-        // 카드 하단의 콘텐츠 기준 y 추적 (가림 판단용)
-        .onGeometryChange(for: CGFloat.self) { proxy in
-            proxy.frame(in: .named(albumDetailScrollSpace)).maxY
-        } action: { newValue in
-            bottomTracker.contentMaxY = newValue
-        }
-        .onAppear {
-            guard isLast else { return }
-            Task { await onLastAppear?() }
-        }
-    }
-}
-
-// MARK: - AlbumDetailLogImageSlider
-// 이미지 슬라이더
-
-private struct AlbumDetailLogImageSlider: View {
-    let images: [AlbumLogImage]
-    
-    @State private var currentIndex: Int = 0
-    
-    private enum Constants {
-        static let cornerRadius: CGFloat = 16
-        static let dotSize: CGFloat = 6
-        static let dotSpacing: CGFloat = 6
-        static let indicatorBottomPadding: CGFloat = 10
-        static let indicatorHPadding: CGFloat = 10
-        static let indicatorVPadding: CGFloat = 6
-        static let placeholderIconSize: CGFloat = 36
-        /// 4:3 비율 높이 계산
-        static var sliderHeight: CGFloat {
-            (UIScreen.main.bounds.width - 40) * 3 / 4
-        }
-    }
-    
-    var body: some View {
-        ZStack(alignment: .bottom) {
-            // 이미지 슬라이더
-            TabView(selection: $currentIndex) {
-                ForEach(images.indices, id: \.self) { index in
-                    AsyncImage(url: images[index].imageUrl) { phase in
-                        switch phase {
-                        case .success(let image):
-                            image
-                                .resizable()    /// 비율 유지 채택 시  ->  .scaledToFill()
-                        default:
-                            // 로드 중 / 실패 시 placeholder 표시
-                            ZStack {
-                                Color.secondary.opacity(0.12)
-                                Image(systemName: "photo")
-                                    .font(.system(size: Constants.placeholderIconSize))
-                                    .foregroundStyle(Color("GrayScale/200"))
-                            }
-                        }
-                    }
-                    .frame(maxWidth: .infinity, maxHeight: Constants.sliderHeight)
-                    .clipped()
-                    .tag(index)
-                }
-            }
-            .tabViewStyle(.page(indexDisplayMode: .never))
-            .frame(height: Constants.sliderHeight)
-            .clipShape(RoundedRectangle(cornerRadius: Constants.cornerRadius))
-            
-            // 커스텀 페이지 인디케이터
-            if images.count > 1 {
-                HStack(spacing: Constants.dotSpacing) {
-                    ForEach(images.indices, id: \.self) { index in
-                        Circle()
-                            .frame(width: Constants.dotSize, height: Constants.dotSize)
-                            .foregroundStyle(
-                                index == currentIndex ? Color("appPrimary/400") : Color.white
-                            )
-                        // 페이지컨트롤dot -> pageControl shadow
-                            .appShadow(.pageControl)
-                    }
-                }
-                .padding(.horizontal, Constants.indicatorHPadding)
-                .padding(.vertical, Constants.indicatorVPadding)
-                .padding(.bottom, Constants.indicatorBottomPadding)
-            }
-        }
-        .frame(maxWidth: .infinity)
-        .frame(height: Constants.sliderHeight)
-    }
-}
-
-// MARK: - AlbumDetailLogImageSkeleton
-// 이미지 응답 대기 중인 로그 카드의 슬라이더 자리에 표시하는 shimmer placeholder
-// 슬라이더와 동일 치수로 그려 카드 레이아웃이 흔들리지 않게 함
-
-private struct AlbumDetailLogImageSkeleton: View {
-    private enum Constants {
-        static let cornerRadius: CGFloat = 16
-        static var sliderHeight: CGFloat {
-            (UIScreen.main.bounds.width - 40) * 3 / 4
-        }
-    }
-
-    var body: some View {
-        RoundedRectangle(cornerRadius: Constants.cornerRadius)
-            .fill(Color.secondary.opacity(0.12))
-            .frame(maxWidth: .infinity)
-            .frame(height: Constants.sliderHeight)
-            .shimmering(shape: RoundedRectangle(cornerRadius: Constants.cornerRadius))
-    }
-}
-
-// MARK: - AlbumDetailLogTextSection
-// 텍스트 + 더보기/접기
-
-private struct AlbumDetailLogTextSection: View {
-    let text: String
-    /// 더보기로 펼쳐질 때 호출 
-    var onExpand: (() -> Void)? = nil
-
-    @State private var isExpanded: Bool = false
-    /// GeometryReader로 측정한 실제 콘텐츠 너비
-    @State private var contentWidth: CGFloat = 0
-    
-    private enum Constants {
-        static let fontName: String = "Pretendard-Regular"
-        static let fontSize: CGFloat = 16
-        static let lineLimit: Int = 2
-        static let moreButtonText: String = "더 보기"
-        static let foldButtonText: String = "접기"
-        static let foldSpacer: String = "  "    /// 접기 버튼과 본문 텍스트 사이 공간 예약용
-        static let reservedGap: CGFloat = 8         // "더 보기"/"접기" 버튼과 본문 사이 간격
-        static let buttonTrailingInset: CGFloat = 8 // 버튼과 우측 끝 사이 간격
-        static let animationDuration: CGFloat = 0.2
-    }
-
-    // 접힌 본문(UITextView)과 동일하게 Dynamic Type 비스케일 고정 크기 사용
-    private var textFont: Font { .custom(Constants.fontName, fixedSize: Constants.fontSize) }
-    private let actionColor = Color("GrayScale/400")
-
-    private var uiFont: UIFont {
-        UIFont(name: Constants.fontName, size: Constants.fontSize)
-        ?? UIFont.systemFont(ofSize: Constants.fontSize)
-    }
-    
-    private var textUIColor: UIColor { UIColor(named: "text") ?? .label }
-
-    // "더 보기" 버튼 폭 + 간격 = 마지막 줄 우측에 비워둘 예약 폭
-    private var reservedTextWidth: CGFloat {
-        (Constants.moreButtonText as NSString)
-            .size(withAttributes: [.font: uiFont]).width
-    }
-    // 마지막 줄 우측 공간
-    private var reservedTailWidth: CGFloat {
-        reservedTextWidth + Constants.reservedGap + Constants.buttonTrailingInset
-    }
-    
-    var body: some View {
-        let truncation = LogTextTruncationAnalyzer.analyze(
-            text: text,
-            font: uiFont,
-            width: contentWidth,
-            reservedTailWidth: reservedTailWidth,
-            lineLimit: Constants.lineLimit
-        )
-
-        Group {
-            if isExpanded {
-                expandedContent
-            } else {
-                collapsedContent(truncation: truncation)
-            }
-        }
-        // 콘텐츠 너비 측정
-        .onGeometryChange(for: CGFloat.self) { $0.size.width } action: { w in
-            if abs(contentWidth - w) > 0.5 {
-                DispatchQueue.main.async { contentWidth = w }
-            }
-        }
-    }
-    
-    // 펼친 상태: 접힘과 같은 UITextView 엔진으로 전체 표시 (편집 페이지와 줄바꿈 일치)
-    // 마지막 줄 끝에 투명 "접기" 예약 텍스트로 공간을 비우고 그 위에 접기 버튼 배치
-    private var expandedContent: some View {
-        ZStack(alignment: .bottomTrailing) {
-            CollapsibleLogTextView(
-                text: text,
-                font: uiFont,
-                textColor: textUIColor,
-                lineLimit: 0,
-                reservedTailWidth: 0,
-                showsEllipsis: false,
-                trailingReserveText: Constants.foldSpacer + Constants.foldButtonText
-            )
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            Button(Constants.foldButtonText) {
-                withAnimation(.easeInOut(duration: Constants.animationDuration)) { isExpanded = false }
-            }
-            .font(textFont)
-            .foregroundStyle(actionColor)
-            .buttonStyle(.plain)
-            .padding(.trailing, Constants.buttonTrailingInset)
-        }
-    }
-    
-    private func expand() {
-        withAnimation(.easeInOut(duration: Constants.animationDuration)) { isExpanded = true }
-        onExpand?()
-    }
-
-    // 접힌 상태: 2줄로 접고, 잘릴 때만 우측에 "더 보기" 버튼 표시
-    private func collapsedContent(truncation: LogTextTruncationAnalyzer.Result) -> some View {
-        ZStack(alignment: .bottomTrailing) {
-            CollapsibleLogTextView(
-                text: text,
-                font: uiFont,
-                textColor: textUIColor,
-                lineLimit: Constants.lineLimit,
-                reservedTailWidth: truncation.isTruncated ? reservedTailWidth : 0,
-                showsEllipsis: truncation.isTruncated && !truncation.lastVisibleLineEndsWithNewline
-            )
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .contentShape(Rectangle())
-            .onTapGesture {
-                guard truncation.isTruncated else { return }
-                expand()
-            }
-
-            if truncation.isTruncated {
-                Button(Constants.moreButtonText) {
-                    expand()
-                }
-                .font(textFont)
-                .foregroundStyle(actionColor)
-                .buttonStyle(.plain)
-                .frame(width: reservedTextWidth, alignment: .trailing)
-                .padding(.trailing, Constants.buttonTrailingInset)
-            }
-        }
     }
 }
